requirements.txt
python-telegram-bot==13
pyowm
python-dotenv==0.19.2